# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "singnalX API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/integrations/telegram/chats": {
      "post": {
        "tags": [
          "Telegram user actions"
        ],
        "summary": "Map Chat",
        "operationId": "map_chat_api_v1_integrations_telegram_chats_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMappingInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMappingOut"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "BackendSession": []
          },
          {
            "BrowserSession": []
          }
        ]
      }
    },
    "/api/v1/integrations/telegram/leads": {
      "get": {
        "tags": [
          "Telegram user actions"
        ],
        "summary": "Leads",
        "operationId": "leads_api_v1_integrations_telegram_leads_get",
        "security": [
          {
            "BackendSession": []
          },
          {
            "BrowserSession": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramLeadPage"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/leads/{lead_id}/telegram": {
      "get": {
        "tags": [
          "Telegram user actions"
        ],
        "summary": "Lead",
        "operationId": "lead_api_v1_leads__lead_id__telegram_get",
        "security": [
          {
            "BackendSession": []
          },
          {
            "BrowserSession": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramLeadOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/leads/{lead_id}/telegram/suggested-reply": {
      "post": {
        "tags": [
          "Telegram user actions"
        ],
        "summary": "Suggested Reply",
        "operationId": "suggested_reply_api_v1_leads__lead_id__telegram_suggested_reply_post",
        "security": [
          {
            "BackendSession": []
          },
          {
            "BrowserSession": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lead Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramLeadOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/api/v1/leads/{lead_id}/telegram/reply": {
      "post": {
        "tags": [
          "Telegram user actions"
        ],
        "summary": "Approved Reply",
        "operationId": "approved_reply_api_v1_leads__lead_id__telegram_reply_post",
        "security": [
          {
            "BackendSession": []
          },
          {
            "BrowserSession": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Lead Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovedReply"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TelegramLeadOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    },
    "/integrations/telegram/webhook": {
      "post": {
        "tags": [
          "Telegram webhook"
        ],
        "summary": "Webhook",
        "operationId": "webhook_integrations_telegram_webhook_post",
        "parameters": [
          {
            "name": "X-Telegram-Bot-Api-Secret-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Telegram-Bot-Api-Secret-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentOutput": {
        "properties": {
          "screening": {
            "$ref": "#/components/schemas/ScreeningResult"
          },
          "qualification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QualificationResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "scoring": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScoringResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "usage": {
            "items": {
              "$ref": "#/components/schemas/UsageInfo"
            },
            "type": "array",
            "title": "Usage"
          },
          "suggested_reply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Reply"
          },
          "decision_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Reason"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "scoring_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scoring Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "screening"
        ],
        "title": "AgentOutput"
      },
...
      "ApprovedReply": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "ApprovedReply"
      },
...
      "ChatMappingInput": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "telegram_chat_id": {
            "type": "integer",
            "exclusiveMaximum": 0.0,
            "exclusiveMinimum": -9.223372036854776e+18,
            "title": "Telegram Chat Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "product_id",
          "telegram_chat_id"
        ],
        "title": "ChatMappingInput"
      },
      "ChatMappingOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "telegram_chat_id": {
            "type": "integer",
            "title": "Telegram Chat Id"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "product_id",
          "telegram_chat_id",
          "enabled"
        ],
        "title": "ChatMappingOut"
      },
      "ContextMessage": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "content": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Content"
          },
          "author": {
            "type": "string",
            "minLength": 1,
            "title": "Author"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "content",
          "author",
          "timestamp"
        ],
        "title": "ContextMessage",
        "description": "Caller supplies only messages from the target's batch/conversation."
      },
...
      "Decision": {
        "type": "string",
        "enum": [
          "IGNORE",
          "REVIEW",
          "RESPOND"
        ],
        "title": "Decision"
      },
      "DeliveryOut": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "not_sent",
              "sending",
              "sent",
              "failed"
            ],
            "title": "Status",
            "default": "not_sent"
          },
          "telegram_message_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Message Id"
          },
          "failure_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Category"
          },
          "delivery_uncertain": {
            "type": "boolean",
            "title": "Delivery Uncertain",
            "default": false
          },
          "draft_busy": {
            "type": "boolean",
            "title": "Draft Busy",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DeliveryOut"
      },
      "DraftRequest": {
        "properties": {
          "regenerate": {
            "type": "boolean",
            "title": "Regenerate",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DraftRequest"
      },
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorOut": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorOut"
      },
...
      "ScoringResult": {
        "properties": {
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "decision": {
            "$ref": "#/components/schemas/Decision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "score",
          "decision"
        ],
        "title": "ScoringResult"
      },
      "ScreeningResult": {
        "properties": {
          "is_candidate": {
            "type": "boolean",
            "title": "Is Candidate"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "is_candidate",
          "reason"
        ],
        "title": "ScreeningResult"
      },
...
      "TelegramLeadOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "message_id": {
            "type": "string",
            "format": "uuid",
            "title": "Message Id"
          },
          "source": {
            "type": "string",
            "const": "telegram",
            "title": "Source",
            "default": "telegram"
          },
          "original_message": {
            "$ref": "#/components/schemas/TelegramMessage"
          },
          "context": {
            "items": {
              "$ref": "#/components/schemas/ContextMessage"
            },
            "type": "array",
            "title": "Context"
          },
          "analysis": {
            "$ref": "#/components/schemas/AgentOutput"
          },
          "delivery": {
            "$ref": "#/components/schemas/DeliveryOut"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "run_id",
          "product_id",
          "message_id",
          "original_message",
          "analysis",
          "delivery"
        ],
        "title": "TelegramLeadOut"
      },
      "TelegramLeadPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TelegramLeadOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "TelegramLeadPage"
      },
      "TelegramMessage": {
        "properties": {
          "source": {
            "type": "string",
            "const": "telegram",
            "title": "Source",
            "default": "telegram"
          },
          "update_id": {
            "type": "integer",
            "exclusiveMaximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Update Id"
          },
          "chat_id": {
            "type": "integer",
            "exclusiveMaximum": 0.0,
            "exclusiveMinimum": -9.223372036854776e+18,
            "title": "Chat Id"
          },
          "message_id": {
            "type": "integer",
            "exclusiveMaximum": 9.223372036854776e+18,
            "exclusiveMinimum": 0.0,
            "title": "Message Id"
          },
          "chat_type": {
            "type": "string",
            "enum": [
              "group",
              "supergroup"
            ],
            "title": "Chat Type"
          },
          "chat_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Title"
          },
          "sender_id": {
            "type": "integer",
            "exclusiveMaximum": 9.223372036854776e+18,
            "exclusiveMinimum": 0.0,
            "title": "Sender Id"
          },
          "sender_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Username"
          },
          "sender_display_name": {
            "type": "string",
            "title": "Sender Display Name"
          },
          "text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "reply_to_message_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMaximum": 9.223372036854776e+18,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply To Message Id"
          },
          "message_thread_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMaximum": 9.223372036854776e+18,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Thread Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "update_id",
          "chat_id",
          "message_id",
          "chat_type",
          "sender_id",
          "sender_display_name",
          "text",
          "timestamp"
        ],
        "title": "TelegramMessage"
      },
...
      "WebhookOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ignored",
              "unmapped",
              "queued",
              "duplicate"
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "WebhookOut"
      }
    },
    "securitySchemes": {
      "BackendSession": {
        "type": "http",
        "scheme": "bearer"
      },
      "BrowserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "signalx_session"
      }
    }
  }
}